Tech-Tards Validator/requirements.txt
streamlit
openai
requests
python-dotenv
pandas
google-generativeai
